use web component for vacation type select element
