+++ src/main/javascript/bundles/app-form.js
@@ -3,10 +3,9 @@ import "../components/timepicker";
 
 import "../js/application/app-form";
 import "../js/application/app-form-day-length-change-listener";
+import "../js/application/vacation-type-select";
 import "../components/back-button";
 import "../components/application-replacement-select";
-
-import "../js/application/vacation-type-changed";
 
 import holidayReplacementChanged from "../js/application/holiday-replacement-changed";
