--- src/main/javascript/js/application/vacation-type-changed.js
@@ -1,23 +1,0 @@
-import $ from "jquery";
-
-$(document).ready(async function () {
-  document.querySelector("#vacationType").addEventListener("change", (event) => {
-    event.preventDefault();
-
-    const value = event.currentTarget.selectedOptions[0].dataset.vacationtypeCategory;
-
-    const overtime = document.querySelector("#overtime");
-    const specialLeave = document.querySelector("#special-leave");
-
-    if (value === "SPECIALLEAVE") {
-      overtime?.classList.add("hidden");
-      specialLeave?.classList.remove("hidden");
-    } else if (value === "OVERTIME") {
-      overtime?.classList.remove("hidden");
-      specialLeave?.classList.add("hidden");
-    } else {
-      overtime?.classList.add("hidden");
-      specialLeave?.classList.add("hidden");
-    }
-  });
-});
