+++ src/main/webapp/WEB-INF/assets-manifest.json
@@ -9,7 +9,7 @@
   "account_form~app_form~app_statistics~overtime_form~sick_note_form~sick_notes~workingtime_form.js": "/assets/account_form~app_form~app_statistics~overtime_form~sick_note_form~sick_notes~workingtime_form.4e60fdb301a14659821d.min.js",
   "app_detail.js": "/assets/app_detail.8f4bb990436d15675d76.min.js",
   "app_detail~app_form~person_overview.js": "/assets/app_detail~app_form~person_overview.791d286f68ff36444f1a.min.js",
-  "app_form.js": "/assets/app_form.0767e99efa094dca7b8c.min.js",
+  "app_form.js": "/assets/app_form.20163a15b5f108f71fa3.min.js",
   "app_list.js": "/assets/app_list.8e1dbd46ee41c1a91558.min.js",
   "app_statistics.js": "/assets/app_statistics.a8884900962436148cd5.min.js",
   "common.css": "/assets/common.7930807cdb48e9d06efd.css",
